Move sharing panel classes to Commons package
