+++ de-lib/src/main/java/org/iplantc/de/analysis/client/gin/factory/AnalysisSharingPresenterFactory.java
@@ -1,7 +1,7 @@
 package org.iplantc.de.analysis.client.gin.factory;
 
 import org.iplantc.de.client.models.analysis.Analysis;
-import org.iplantc.de.client.sharing.SharingPresenter;
+import org.iplantc.de.commons.client.presenter.SharingPresenter;
 
 import java.util.List;
