+++ de-lib/src/main/java/org/iplantc/de/apps/client/gin/factory/AppSharingPresenterFactory.java
@@ -1,7 +1,7 @@
 package org.iplantc.de.apps.client.gin.factory;
 
 import org.iplantc.de.client.models.apps.App;
-import org.iplantc.de.client.sharing.SharingPresenter;
+import org.iplantc.de.commons.client.presenter.SharingPresenter;
 
 import java.util.List;
