+++ de-lib/src/main/java/org/iplantc/de/apps/client/views/sharing/dialog/AppSharingDialog.java
@@ -6,7 +6,7 @@ package org.iplantc.de.apps.client.views.sharing.dialog;
 
 import org.iplantc.de.apps.client.gin.factory.AppSharingPresenterFactory;
 import org.iplantc.de.client.models.apps.App;
-import org.iplantc.de.client.sharing.SharingPresenter;
+import org.iplantc.de.commons.client.presenter.SharingPresenter;
 import org.iplantc.de.commons.client.views.dialogs.IPlantDialog;
 
 import com.google.common.base.Preconditions;
