+++ de-lib/src/main/java/org/iplantc/de/commons/client/gin/CommonsGinModule.java
@@ -1,8 +1,8 @@
-package org.iplantc.de.client.gin;
+package org.iplantc.de.commons.client.gin;
 
-import org.iplantc.de.client.gin.factory.SharingPermissionViewFactory;
-import org.iplantc.de.client.sharing.SharingPermissionView;
-import org.iplantc.de.client.sharing.SharingPermissionsPanel;
+import org.iplantc.de.commons.client.gin.factory.SharingPermissionViewFactory;
+import org.iplantc.de.commons.client.views.sharing.SharingPermissionView;
+import org.iplantc.de.commons.client.views.sharing.SharingPermissionsPanel;
 
 import com.google.gwt.inject.client.AbstractGinModule;
 import com.google.gwt.inject.client.assistedinject.GinFactoryModuleBuilder;
@@ -10,7 +10,7 @@ import com.google.gwt.inject.client.assistedinject.GinFactoryModuleBuilder;
 /**
  * @author aramsey
  */
-public class ClientGinModule extends AbstractGinModule {
+public class CommonsGinModule extends AbstractGinModule {
     @Override
     protected void configure() {
         install(new GinFactoryModuleBuilder().implement(SharingPermissionView.class,
