+++ de-lib/src/main/java/org/iplantc/de/commons/client/gin/factory/SharingPermissionViewFactory.java
@@ -1,8 +1,8 @@
-package org.iplantc.de.client.gin.factory;
+package org.iplantc.de.commons.client.gin.factory;
 
 import org.iplantc.de.client.models.sharing.SharedResource;
-import org.iplantc.de.client.sharing.SharingPermissionView;
-import org.iplantc.de.client.sharing.SharingPresenter;
+import org.iplantc.de.commons.client.views.sharing.SharingPermissionView;
+import org.iplantc.de.commons.client.presenter.SharingPresenter;
 
 import com.sencha.gxt.core.shared.FastMap;
