+++ de-lib/src/main/java/org/iplantc/de/commons/client/presenter/SharingPresenter.java
@@ -1,4 +1,4 @@
-package org.iplantc.de.client.sharing;
+package org.iplantc.de.commons.client.presenter;
 
 import org.iplantc.de.client.models.diskResources.PermissionValue;
