+++ de-lib/src/main/java/org/iplantc/de/commons/client/views/sharing/SharingPermissionsView.ui.xml
@@ -8,7 +8,7 @@
              xmlns:toolbar="urn:import:com.sencha.gxt.widget.core.client.toolbar">
 
     <ui:with field="appearance"
-             type="org.iplantc.de.client.sharing.SharingAppearance"/>
+             type="org.iplantc.de.commons.client.views.sharing.SharingAppearance"/>
     <ui:with field="listStore"
              type="com.sencha.gxt.data.shared.ListStore"/>
     <ui:with field="cm"
