+++ de-lib/src/main/java/org/iplantc/de/diskResource/client/gin/factory/DataSharingPresenterFactory.java
@@ -1,7 +1,7 @@
 package org.iplantc.de.diskResource.client.gin.factory;
 
 import org.iplantc.de.client.models.diskResources.DiskResource;
-import org.iplantc.de.client.sharing.SharingPresenter;
+import org.iplantc.de.commons.client.presenter.SharingPresenter;
 
 import java.util.List;
