+++ de-lib/src/main/java/org/iplantc/de/diskResource/client/views/sharing/DataSharingViewImpl.java
@@ -1,8 +1,8 @@
 package org.iplantc.de.diskResource.client.views.sharing;
 
 import org.iplantc.de.client.models.diskResources.DiskResource;
-import org.iplantc.de.client.sharing.SharingAppearance;
-import org.iplantc.de.client.sharing.SharingPresenter;
+import org.iplantc.de.commons.client.views.sharing.SharingAppearance;
+import org.iplantc.de.commons.client.presenter.SharingPresenter;
 import org.iplantc.de.client.util.DiskResourceUtil;
 import org.iplantc.de.diskResource.client.DataSharingView;
 import org.iplantc.de.diskResource.client.model.DiskResourceModelKeyProvider;
