+++ de-lib/src/main/java/org/iplantc/de/diskResource/client/views/sharing/dialogs/ShareBreakDownDialog.java
@@ -2,7 +2,7 @@ package org.iplantc.de.diskResource.client.views.sharing.dialogs;
 
 import org.iplantc.de.client.models.diskResources.PermissionValue;
 import org.iplantc.de.client.models.sharing.Sharing;
-import org.iplantc.de.client.sharing.SharingAppearance;
+import org.iplantc.de.commons.client.views.sharing.SharingAppearance;
 import org.iplantc.de.diskResource.client.model.DataSharingKeyProvider;
 import org.iplantc.de.diskResource.client.model.DataSharingProperties;
