+++ de-webapp/src/main/java/org/iplantc/de/admin/client/gin/BelphegorAppInjector.java
@@ -6,7 +6,7 @@ import org.iplantc.de.admin.desktop.client.ontologies.gin.OntologiesGinModule;
 import org.iplantc.de.admin.desktop.client.toolAdmin.gin.ToolAdminGinModule;
 import org.iplantc.de.admin.desktop.client.views.BelphegorView;
 import org.iplantc.de.admin.desktop.client.workshopAdmin.gin.WorkshopAdminGinModule;
-import org.iplantc.de.client.gin.ClientGinModule;
+import org.iplantc.de.commons.client.gin.CommonsGinModule;
 import org.iplantc.de.collaborators.client.gin.CollaboratorsGinModule;
 import org.iplantc.de.commons.client.comments.gin.CommentsGinModule;
 import org.iplantc.de.diskResource.client.gin.DiskResourceGinModule;
@@ -25,7 +25,7 @@ import com.google.gwt.inject.client.Ginjector;
               AdminAppsGinModule.class,
               OntologiesGinModule.class,
               DiskResourceGinModule.class,
-              ClientGinModule.class,
+              CommonsGinModule.class,
               CollaboratorsGinModule.class,
               CommentsGinModule.class,
               TagsGinModule.class,
